Update sessions and registrations controllers location, update routes
