--- app/controllers/static_controller.rb
@@ -1,5 +1,0 @@
-class StaticController < ApplicationController
-  def home
-    render json: { status: 'success' }
-  end
-end
